Two changes made to the TodoApp and TodoAdd
I made two changes and updates to the two components which are TodoApp
and TodoAdd. TodoApp now contains a new state, previous commit didn't
have that. Now we will store todoArray property in the state which will
contain an array consisting of object helpful for rendering Todo
components

As for the TodoAdd component change made is new ref property provided to
the input[type="text"] element inside the render function. It fixed the
problem of inputRef.current is null.